% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_monthly_covariates.R
\name{add_monthly_covariates}
\alias{add_monthly_covariates}
\title{Function to add covariate data to a dataframe}
\usage{
add_monthly_covariates(target_df, monthly_data)
}
\arguments{
\item{target_df}{Target dataframe with day-level observations, where "date" is the last column}

\item{monthly_data}{Dataframe of covariates, recorded on a monthly basis on the 1st day of each month, where "date" is the last column}
}
\description{
This function performs a merge of monthly covariate data into a dataframe with day-level observations. E.g. I have multiple observations from January 2012 and I want to add the monthly January covariate data to each of those observations.
}
\examples{
add_monthly_covariates(FARC_results, monthly_viol)
}
\keyword{monthly_covariates}
