% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_LIWC_values.R
\name{extract_liwc_values}
\alias{extract_liwc_values}
\title{Function to get raw LIWC measures for a corpus of text}
\usage{
extract_liwc_values(df, LIWC_dict)
}
\arguments{
\item{df}{Dataframe where text is in column "text" and date is in column "date"}

\item{LIWC_dict}{Name of LIWC dictionary to use}
}
\description{
This function gets LIWC values for a corpus of text and returns them in a formatted dataframe.
}
\examples{
extract_liwc_values(FARC, spanish_dict)
}
\keyword{LIWC}
