% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_base_actor.R
\name{generate_base_actor}
\alias{generate_base_actor}
\title{Function to generate a base graph by actor}
\usage{
generate_base_actor(data, title)
}
\arguments{
\item{data}{Dataframe of LIWC estimates}

\item{title}{Title for the graph}
}
\description{
This function generates a time series graph for 1 actor
}
\examples{
generate_base_actor(FARC_results, "FARC: Loessed \% of Document Contributed by Sentiment Type")
}
\keyword{ggplot,}
\keyword{graph}
