% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_binary_states.R
\name{generate_binary_states}
\alias{generate_binary_states}
\title{Function to calculate binary sentiment level (low or high)}
\usage{
generate_binary_states(df, side)
}
\arguments{
\item{df}{Dataframe of LIWC estimates}

\item{side}{Side string}
}
\description{
This function takes LIWC point estimates (can be loessed) based on a measure from Pennebaker et al.
Sentiment = high if 2 * % negative emotion is less than or equal to % positive emotion
}
\examples{
generate_binary_states(df, side)
}
\keyword{binary_sentiment}
