% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_add_dates.R
\name{graph_add_dates}
\alias{graph_add_dates}
\title{Function to add major dates to a base graph}
\usage{
graph_add_dates(base_graph, title, data, group_label, legend_label)
}
\arguments{
\item{base_graph}{Ggplot graph to which datelines will be added}

\item{title}{Title for the graph}

\item{data}{Dataframe of dates with categorical labels}

\item{group_label}{Group i.e. categorical variable value to graph}

\item{legend_label}{Name of group for the graph legend}
}
\description{
This function adds major agreements to a base graph
}
\examples{
graph_add_dates(negative_sentiment_graph, "Negative Sentiment and Major Agreements", dates, "major_agree", "Major agreements")
}
\keyword{ggplot,}
\keyword{graph}
