% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loess_liwc_estimates.R
\name{loess_liwc_estimates}
\alias{loess_liwc_estimates}
\title{Function to calculate loess point estimates of LIWC estimates}
\usage{
loess_liwc_estimates(liwc_results)
}
\arguments{
\item{liwc_results}{Dataframe of LIWC values, where "date" is the last observation}
}
\description{
This function returns a dataframe of loessed point estimates of raw LIWC measures.
}
\examples{
loess_liwc_estimates(FARC_raw)

}
\keyword{LIWC,}
\keyword{loess}
