% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_base_graph.R
\name{generate_base_graph}
\alias{generate_base_graph}
\title{Function to generate a base graph of sentiment}
\usage{
generate_base_graph(y_var, y_legend, title)
}
\arguments{
\item{y_var}{Name of the y variable to be plotted (must match column name in the dataframes)}

\item{y_legend}{Name for the y variable to appear in the legend}

\item{title}{Title for the graph}
}
\description{
This function generates a time series graph of 1 type of sentiment x the 3 actors. Includes scatterplot of observed values as well as loess curves.
}
\keyword{ggplot,}
\keyword{graph}
