% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liwc_summary_stats.R
\name{liwc_summary_stats}
\alias{liwc_summary_stats}
\title{Function to generate summary statistics from a LIWC matrix}
\usage{
liwc_summary_stats(df)
}
\arguments{
\item{df}{Dataframe of texts}
}
\description{
Function to generate summary statistics from a LIWC matrix
}
\examples{
liwc_summary_stats(FARC)
}
\keyword{statistics}
\keyword{summary}
